+++ Changes.md
@@ -2,8 +2,9 @@
 
 ## Version 1.5
 
-- Shows alert state (red/yellow flashing) on FSD buttons when active but also
-  overheating
+- Shows active state (yellow) on silent running buttons
+- Shows alert state (red/yellow flashing) on FSD and silent running buttons if
+  that control is active but ship is also overheating
 - Shows correct states immediately, i.e. do not wait for first significant state
   change before updating.
 - Shows correct states consitently on buttons where one LED represents multiple